refactor(state): move update_starknet_state to pathfinder_merkle_tree
