+++ crates/pathfinder/src/state.rs
@@ -1,13 +1,4 @@
 pub mod block_hash;
 mod sync;
 
-pub use sync::{
-    l1,
-    l2,
-    revert,
-    sync,
-    update_starknet_state,
-    Gossiper,
-    SyncContext,
-    RESET_DELAY_ON_FAILURE,
-};
+pub use sync::{l1, l2, revert, sync, Gossiper, SyncContext, RESET_DELAY_ON_FAILURE};
